+++ test/main.tf
@@ -1,7 +1,7 @@
 terraform {
   required_providers {
     gopackager = {
-      version = "0.3.0"
+      version = "0.4.0"
       source  = "github.com/stevencyb/gopackager"
     }
   }
@@ -13,6 +13,16 @@ data "gopackager_compile" "example_local" {
   destination = "build/a/bootstrap"
   goarch      = "amd64"
   goos        = "linux"
+}
+
+data "gopackager_compile" "example_local_git" {
+  source      = "../main.go"
+  destination = "build/a/bootstrap"
+  goarch      = "amd64"
+  goos        = "linux"
+
+  git_trigger      = true
+  git_trigger_path = "../internal"
 }
 
 data "gopackager_compile" "example_local_zip" {
@@ -37,6 +47,13 @@ output "example_local" {
     output_sha512        = data.gopackager_compile.example_local.output_sha512
     output_sha256_base64 = data.gopackager_compile.example_local.output_sha256_base64
     output_sha512_base64 = data.gopackager_compile.example_local.output_sha512_base64
+  }
+}
+
+output "example_local_git" {
+  value = {
+    output_path     = data.gopackager_compile.example_local_git.output_path
+    output_git_hash = data.gopackager_compile.example_local_git.output_git_hash
   }
 }